Added tests for Mvc\Micro\Collection.
