+++ tests/integration/Mvc/Micro/Collection/MapViaCest.php
@@ -13,6 +13,9 @@ declare(strict_types=1);
 namespace Phalcon\Test\Integration\Mvc\Micro\Collection;
 
 use IntegrationTester;
+use Phalcon\Mvc\Micro;
+use Phalcon\Mvc\Micro\Collection;
+use Phalcon\Test\Controllers\MicroController;
 
 /**
  * Class MapViaCest
@@ -23,13 +26,28 @@ class MapViaCest
      * Tests Phalcon\Mvc\Micro\Collection :: mapVia()
      *
      * @param IntegrationTester $I
-     *
-     * @author Phalcon Team <team@phalconphp.com>
-     * @since  2018-11-13
      */
-    public function mvcMicroCollectionMapVia(IntegrationTester $I)
+    public function testMicroCollectionVia(IntegrationTester $I)
     {
         $I->wantToTest('Mvc\Micro\Collection - mapVia()');
-        $I->skipTest('Need implementation');
+
+        $app        = new Micro();
+        $collection = new Collection();
+
+        $collection->setHandler(new MicroController());
+
+        $collection->mapVia(
+            "/test",
+            'indexAction',
+            ["POST", "GET"],
+            "test"
+        );
+
+        $app->mount($collection);
+
+        $expected = ["POST", "GET"];
+        $actual   = $app->getRouter()->getRouteByName("test")->getHttpMethods();
+
+        $I->assertEquals($expected, $actual);
     }
 }
