+++ tests/integration/Mvc/Micro/Collection/SetHandlerCest.php
@@ -13,6 +13,9 @@ declare(strict_types=1);
 namespace Phalcon\Test\Integration\Mvc\Micro\Collection;
 
 use IntegrationTester;
+use Phalcon\Mvc\Micro;
+use Phalcon\Mvc\Micro\Collection;
+use Phalcon\Test\Controllers\Micro\Collections\PersonasLazyController;
 
 /**
  * Class SetHandlerCest
@@ -31,5 +34,40 @@ class SetHandlerCest
     {
         $I->wantToTest('Mvc\Micro\Collection - setHandler()');
         $I->skipTest('Need implementation');
+    }
+
+    public function testMicroCollectionsLazy(IntegrationTester $I)
+    {
+        $app        = new Micro();
+        $collection = new Collection();
+
+        $collection->setHandler(
+            PersonasLazyController::class,
+            true
+        );
+
+
+        $collection->map('/', 'index');
+        $collection->map('/edit/{number}', 'edit');
+
+        $app->mount($collection);
+
+
+
+        $app->handle('/');
+
+        $I->assertEquals(
+            1,
+            PersonasLazyController::getEntered()
+        );
+
+
+
+        $app->handle('/edit/100');
+
+        $I->assertEquals(
+            101,
+            PersonasLazyController::getEntered()
+        );
     }
 }
