+++ tests/integration/Mvc/Refactor-MicroCest.php
@@ -16,7 +16,6 @@ use Phalcon\Di\FactoryDefault;
 use Phalcon\Events\Event;
 use Phalcon\Events\Manager;
 use Phalcon\Mvc\Micro;
-use Phalcon\Test\Controllers\MicroController;
 use Phalcon\Test\Fixtures\Micro\MyMiddleware;
 use Phalcon\Test\Fixtures\Micro\MyMiddlewareStop;
 use Phalcon\Test\Fixtures\Micro\RestHandler;
@@ -411,24 +410,6 @@ class MicroCest
 
         $expected = 'success';
         $actual   = $app->handle('/api');
-        $I->assertEquals($expected, $actual);
-    }
-
-    public function testMicroCollectionVia(IntegrationTester $I)
-    {
-        $app        = new Micro();
-        $collection = new Micro\Collection();
-        $collection->setHandler(new MicroController());
-        $collection->mapVia(
-            "/test",
-            'indexAction',
-            ["POST", "GET"],
-            "test"
-        );
-        $app->mount($collection);
-
-        $expected = ["POST", "GET"];
-        $actual   = $app->getRouter()->getRouteByName("test")->getHttpMethods();
         $I->assertEquals($expected, $actual);
     }
 }
